Refactor login component: replace Login with AuthForm and update navigation links
