+++ frontend/src/app/login/page.tsx
@@ -1,9 +1,9 @@
-import Login from "@/components/Login";
+import AuthForm from "@/components/AuthForm";
 
 export default function Landing() {
   return (
     <section className="overflow-x-hidden flex justify-center items-center h-dvh w-dvw">
-        <Login />
+        <AuthForm />
     </section>
   );
 }
