+++ frontend/src/components/Navabar.tsx
@@ -13,7 +13,7 @@ const Navbar: React.FC = () => {
             </div>
             <ul className="flex flex-row justify-center gap-2 xs:gap-4 items-center">
                 <li>
-                    <Link href="/" className='text-[1.2rem]'>login</Link>
+                    <Link href="/login" className='text-[1.2rem]'>login</Link>
                 </li>
                 <li className='hidden xs:block h-2 w-2 bg-dark-green rounded'></li>
                 <li className="px-5 xs:px-10 border-2 py-1 border-dark-green hover:text-light-green cursor-pointer relative overflow-hidden transition-all duration-700 active:duration-500 active:bg-[#1a393d] active:border-[#1a393d] before:content-[''] before:absolute before:bg-dark-green before:transition-all before:duration-700 before:w-[200%] before:h-[200%] before:top-[110%] before:left-[110%]  hover:before:top-[-30px] hover:before:left-[-30px] before:z-[-1] before:rounded">
